LayoutTests/http/tests/privateClickMeasurement/attribution-conversion-through-image-redirect-in-new-window.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html lang="en">
<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
    <script src="/js-test-resources/ui-helper.js"></script>
    <script src="resources/util.js"></script>
</head>
<body onload="setTimeout(runTest, 0)">
<div id="description">Tests triggering of private click measurement attributions in a new window.</div>
<a target="_blank" rel="opener" id="targetLink" href="http://localhost:8000/privateClickMeasurement/resources/convertAndPostMessageBack.html" attributionsourceid=3 attributiondestination="http://localhost:8000">Link</a><br>
<div id="output"></div>
<script>
    prepareTest();

    function activateElement(elementID) {
        var element = document.getElementById(elementID);
        var centerX = element.offsetLeft + element.offsetWidth / 2;
        var centerY = element.offsetTop + element.offsetHeight / 2;
        UIHelper.activateAt(centerX, centerY).then(
            function () {
            },
            function () {
                document.getElementById("output").innerText = "FAIL Promise rejected.";
                tearDownAndFinish();
            }
        );
    }

    function receiveMessage(event) {
        if (event.origin === "http://localhost:8000") {
            if (event.data === "Done") {
                testRunner.dumpPrivateClickMeasurement();
                document.body.removeChild(document.getElementById("targetLink"));
                tearDownAndFinish();
            } else {
                testFailed("Received unknown message: " + event.data);
                tearDownAndFinish();
            }
        } else {
            testFailed("Received a message from an unexpected origin: " + event.origin);
            tearDownAndFinish();
        }
    }

    window.addEventListener("message", receiveMessage, false);

    function runTest() {
        if (window.testRunner) {
            testRunner.setCanOpenWindows();
            activateElement("targetLink");
        } else {
            document.getElementById("output").innerText = "FAIL No testRunner.";
            tearDownAndFinish();
        }
    }
</script>
</body>
</html>
