LayoutTests/http/tests/privateClickMeasurement/attribution-conversion-through-image-redirect-in-new-window-expected.txt

Tests triggering of private click measurement conversions in a new window.


Converted Private Click Measurements:
WebCore::PrivateClickMeasurement 1
Source: 127.0.0.1
Destination: localhost
Campaign ID: 3
Conversion data: 12
Conversion priority: 3
Conversion earliest time to send: Within 24-48 hours
Conversion request sent: false
